requirements.txt
cloudify-plugins-common==3.0
docker-py==0.4.0
netifaces==0.10.4
